refactor(web): split ai-suggestions host
Extract the message-thread reducer (messages, pending accepts, stale
recompute) into useAISuggestionThread. Extract the one-shot attention
flash pattern (setAttention + setTimeout) into a shared usePulse hook;
migrate the inspector facet-bar pulse to the same hook.
